PCI: Make sure bridge windows include their fixed BARs
When choosing a start address for a bridge window, it should be not just a
lowest possible address: this window must cover every underlying fixed BAR.
The lowest address that satisfies this requirement is the .realloc_range
field of struct pci_bus.

After allocating a bridge window, validate that it covers all its fixed
BARs: this range is put to the .fixed_range field of struct pci_bus.

Signed-off-by: Sergei Miroshnichenko <s.miroshnichenko@yadro.com>