scsi: qla2xxx: Reduce timeout if cable is unplugged
Target creation (i.e. write of "1" to "tpg_1/enable") blocks for 23
seconds if cable is not connected. Firmware always returns "loss of
sync" state during the period. It slows down management software.

If there's a cable, firmware may not report "loss of sync" state at all.
E.g. in the initiator mode, the first state reported after driver
initialization may be "wait login". Then, within 500 ms, it reaches
"ready" state. In pure target mode it was observed that firmware goes
directly to the "ready" state if there's cable.

Therefore it should be safe to unblock configfs much earlier if firmware
stays in "loss of sync" only for 5 seconds. Total execution time of the
write to "tpg_1/enable" is ~4 seconds if there's a cable and ~9 seconds
otherwise.

Signed-off-by: Ivan Efremov <i.efremov@yadro.com>
Signed-off-by: Roman Bolshakov <r.bolshakov@yadro.com>