scsi: mpt3sas: Handle a surprise PCI unplug faster
When a HBA disappears from the PCI bus (due to a surprise unplug or a h/w
failure), unloading its driver may take tens of minutes, and the lsscsi
still shows the already-removed devices before that finishes. This happens
because the SYNCHRONIZE_CACHE is queued and allowed during driver unload,
waiting for a 1 minute timeout for each drive.

Add a new dead_host flag, accompanying the remove_host flag, to indicate an
inability to queue any command. Now a cleanup after a sudden HBA removal
takes a few seconds.

Fixes: 9029a72500b9 ("scsi: mpt3sas: Fix Sync cache command failure during driver unload")
CC: linux-scsi@vger.kernel.org
CC: Suganath Prabu S <suganath-prabu.subramani@broadcom.com>
Signed-off-by: Sergei Miroshnichenko <s.miroshnichenko@yadro.com>