PCI: Movable BARs: Make just a single attempt to assign bridge resources
With enabled BAR movement, all BARs and bridge windows are released at the
beginning of a reassign procedure, not just part of them. In this case only
one pass is enough, no need for several traverses.

In case of failures the pci_reassign_root_bus_resources() disables BARs for
one of the hot-added devices and retries the single-step assignment.

Signed-off-by: Sergei Miroshnichenko <s.miroshnichenko@yadro.com>