PCI: hotplug: Enable the movable BARs feature by default
This is the last patch in the series which implements the essentials of the
movable BARs feature, so it is turned on by default now.

Tested on x86_64; and with extra patches applied, it also works on ppc64le
(PowerNV).

In case of problems it is still can be overridden by the following command
line option:

  pcie_movable_bars=off

Signed-off-by: Sergei Miroshnichenko <s.miroshnichenko@yadro.com>